+++ Source/Gapotchenko.GnuTK/Toolkits/ToolkitServices.cs
@@ -122,9 +122,7 @@ static class ToolkitServices
         return
             portableToolkits // portable toolkits have priority
             .Concat(installedToolkits) // installed toolkits come next            
-            .DistinctBy( // with no duplicates
-                toolkit => (toolkit.Name, toolkit.InstallationPath),
-                ValueTupleEqualityComparer.Create(StringComparer.Ordinal, FileSystem.PathComparer));
+            .DistinctBy(toolkit => toolkit.Name, StringComparer.OrdinalIgnoreCase); // with no duplicates
     }
 
     /// <summary>