+++ Documentation/Changelog.md
@@ -8,6 +8,7 @@ Release date: not released yet
 
 - Use `errexit` and `pipefail` shell options when executing commands
 - Utilize MSYS2 toolkits installed by `msys2/setup-msys2` GitHub action
+- Fixed toolkit deduplication
 
 ### GNU-TK 2025.3